# rule_engine/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "wArmor规则引擎API接口文档",
        "title": "wArmor规则引擎API服务",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/config": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "获取waf当前运行模式",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "修改waf运行模式",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "raw",
                        "name": "raw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "创建waf运行模式",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "raw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Config"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/config/enum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "配置"
                ],
                "summary": "获取waf模式枚举对应关系",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ip": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP黑白名单"
                ],
                "summary": "获取IP黑白名单列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "封禁类型【枚举值】",
                        "name": "block_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "ip类型【枚举值】",
                        "name": "ip_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前页数,默认值:1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前条数,默认值:10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP黑白名单"
                ],
                "summary": "创建IP黑白名单",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "raw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Ip"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP黑白名单"
                ],
                "summary": "删除IP黑白名单",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ip/enum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP黑白名单"
                ],
                "summary": "获取IP黑白名单枚举对应关系",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/ip/remove": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IP黑白名单"
                ],
                "summary": "定时任务检测黑名单IP是否过期,若过期则删除",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/rule": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "规则"
                ],
                "summary": "获取规则列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "模糊查询",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "规则状态【0=false,1=true】",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "规则变量【枚举值】",
                        "name": "rule_variable",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "规则类型【枚举值】",
                        "name": "rule_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "规则动作【枚举值】",
                        "name": "rule_action",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "规则级别【枚举值】",
                        "name": "severity",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前页数,默认值:1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "当前条数,默认值:10",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "规则"
                ],
                "summary": "修改规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "raw",
                        "name": "raw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Rule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "规则"
                ],
                "summary": "创建规则",
                "parameters": [
                    {
                        "description": "raw",
                        "name": "raw",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Rule"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "规则"
                ],
                "summary": "删除规则",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/rule/enum": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "规则"
                ],
                "summary": "获取规则枚举对应关系",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Config": {
            "type": "object",
            "required": [
                "mode",
                "operator"
            ],
            "properties": {
                "mode": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "operator": {
                    "type": "string"
                }
            }
        },
        "model.Ip": {
            "type": "object",
            "required": [
                "comment",
                "ip_address",
                "ip_type",
                "operator"
            ],
            "properties": {
                "block_type": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "comment": {
                    "type": "string"
                },
                "expire_time_tag": {
                    "type": "string",
                    "example": "当创建黑名单IP并且是临时封禁时,1:封禁1小时,2:封禁8小时,3:封禁1天,4:封禁7天。创建黑名单永久封禁以及白名单此字段无需填写"
                },
                "ip_address": {
                    "type": "string"
                },
                "ip_type": {
                    "type": "integer",
                    "enum": [
                        1,
                        2
                    ]
                },
                "operator": {
                    "type": "string"
                }
            }
        },
        "model.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "model.Rule": {
            "type": "object",
            "required": [
                "description",
                "operator",
                "rule_action",
                "rule_type",
                "rule_variable",
                "rules",
                "rules_operation",
                "severity",
                "status"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "operator": {
                    "type": "string"
                },
                "rule_action": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3
                    ]
                },
                "rule_type": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5,
                        6,
                        7,
                        8,
                        9,
                        10,
                        11,
                        12,
                        13,
                        14,
                        15
                    ]
                },
                "rule_variable": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5,
                        6
                    ]
                },
                "rules": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "规则正则表达式列表"
                    ]
                },
                "rules_operation": {
                    "type": "string",
                    "example": "输入 and 或 or,and需全部匹配Rules列表,or匹配Rules列表任意一个"
                },
                "severity": {
                    "type": "integer",
                    "enum": [
                        1,
                        2,
                        3,
                        4,
                        5
                    ]
                },
                "status": {
                    "type": "boolean"
                }
            }
        }
    }
}